// src/types/ProductDescription.ts
export interface ProductDescription {
    id?: string;
    product_id: string;
    general_description: string;
    dm_or_dmd: string;
    dm_class: string;
    product_reference: string;
    on_market: boolean;
    gmn: string;
    gtin: string;
    classification_code: string;
    intended_use: string;
    intended_user: string;
    operating_principle: string;
    medical_indication: string;
    contraindication: string;
    performance_claims: string;
    expected_benefit: string;
}